+++ Tests/UnitTests/Core/Propagator/CMakeLists.txt
@@ -8,7 +8,8 @@ add_unittest(JacobianEngine JacobianEngineTests.cpp)
 add_unittest(KalmanExtrapolator KalmanExtrapolatorTests.cpp)
 add_unittest(LoopProtection LoopProtectionTests.cpp)
 add_unittest(MaterialCollection MaterialCollectionTests.cpp)
-add_unittest(MultiStepper MultiStepperTests.cpp)
+add_unittest(MultiEigenStepperLoop MultiEigenStepperLoopTests.cpp)
+add_unittest(MultiSympyStepperLoop MultiSympyStepperLoopTests.cpp)
 add_unittest(Navigator NavigatorTests.cpp)
 add_unittest(Propagator PropagatorTests.cpp)
 add_unittest(EigenStepper EigenStepperTests.cpp)
